@@ -5,36 +5,32 @@ declare(strict_types=1);
 namespace App\Application\Command\User\SignIn;
 
 use App\Application\Command\CommandHandler;
-use App\Domain\User\Checker\CheckUserByEmail;
+use App\Domain\Shared\Exception\NotFoundException;
 use App\Domain\User\Exception\InvalidCredentialsException;
+use App\Domain\User\Finder\FindUserByEmail;
 use App\Domain\User\Repository\UserRepository;
+use Ramsey\Uuid\Uuid;
 
 final class SignInHandler implements CommandHandler
 {
-    /**
-     * @var UserRepository
-     */
-    private $userStore;
-    /**
-     * @var CheckUserByEmail
-     */
-    private $userChecker;
+    private UserRepository $userStore;
+    private FindUserByEmail $findUserByEmail;
 
-    public function __construct(UserRepository $userStore, CheckUserByEmail $checker)
+    public function __construct(UserRepository $userStore, FindUserByEmail $findUserByEmail)
     {
         $this->userStore = $userStore;
-        $this->userChecker = $checker;
+        $this->findUserByEmail = $findUserByEmail;
     }
 
     public function __invoke(SignInCommand $command)
     {
-        $uuid = $this->userChecker->existsEmail($command->email);
-
-        if (null === $uuid) {
+        try {
+            $userView = $this->findUserByEmail->oneByEmail($command->email);
+        } catch (NotFoundException $e) {
             throw new InvalidCredentialsException();
         }
 
-        $user = $this->userStore->get($uuid);
+        $user = $this->userStore->get(Uuid::fromString($userView->uuid));
 
         $user->signIn($command->plainPassword);
     }
