@@ -12,18 +12,9 @@ use App\Infrastructure\Shared\Flusher\Flusher;
 
 final class SignUpHandler implements CommandHandler
 {
-    /**
-     * @var UserRepository
-     */
-    private $userStore;
-    /**
-     * @var UniqueEmail
-     */
-    private $uniqueEmail;
-    /**
-     * @var Flusher
-     */
-    private $flusher;
+    private UserRepository $userStore;
+    private UniqueEmail $uniqueEmail;
+    private Flusher $flusher;
 
 
     public function __construct(UserRepository $userStore, UniqueEmail $uniqueEmail, Flusher $flusher)
